+++ src/prodigy_prot/models.py
@@ -1,10 +1,3 @@
-#!/usr/bin/env python
-#
-# This code is part of the binding affinity prediction tools distribution
-# and governed by its license.  Please see the LICENSE file that should
-# have been included as part of this package.
-#
-
 """
 Models to predict binding affinity based on molecular properties.
 """
