--- src/prodigy_prot/modules/__init__.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python
-#
-# This code is part of the binding affinity prediction tools distribution
-# and governed by its license.  Please see the LICENSE file that should
-# have been included as part of this package.
-#
-
-"""
-Binding affinity prediction methods developed by the Bonvin Lab.
-"""
